requirements.txt
jupyter
pandas
future
numpy
scipy
sklearn
Pillow
pytest
